requirements.txt
torch==2.0.1
torchvision==0.18.1
efficientnet-pytorch==0.7.1
numpy==1.26.4
opencv-python==4.5.3.56
Pillow==8.3.2
Flask==2.0.2
gunicorn==20.1.0
Werkzeug==2.0.2python.
requests==2.26.0
pandas==2.2.2